Converted the ErrorLoggingDispatcher into an ErrorHandlingDispatcher
